Conflict Finder added, Plan merger modified
Added an Internal and User Conflict Solver.
The Internal is only used by plan mergers.
The User Conflict Solver can be used to identify Conflicts in Plans.
It outputs the type of conflict and other relevant information.

Last but not least modfied the Plan Merger to use the representation of the conflict finder.