Benchmark Revamp + Plan Merger Modifications
Complety revamped the Benchmark File Structure.
Each now has it's own directory as well as a basic Read Me.

Modified the Plan Merger.
There are now 2 Plan Mergers, on working only with the conflict Predicates by the Conflict finder.
The other one is working with the usual occurs of asprilo. (This is currently more fleshed out than the conflict one.)